pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "DuckAI", email = "duckai@gmail.com"},
]
dependencies = [
    "djangorestframework>=3.14.0",
    "setuptools>=69.0.2",
    "django>=4.2.7",
    "djangorestframework>=3.14.0",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
    "djangorestframework-simplejwt>=5.3.0",
    "openai>=1.3.5",
    "drf-spectacular>=0.26.5",
    "python-magic>=0.4.27",
    "Pillow>=10.1.0",
    "replicate>=0.20.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"

[tool.pdm.external]
python = { version = "3.8", markers = "sys_platform == 'linux'", include = ["libmagic1"] }


